+++ R/setup.R
@@ -16,24 +16,24 @@
 #' @export
 #' @rdname setup
 setup_machine <- function() {
-  cli({
-    cli_h1("Machine setup")
+  cli::cli({
+    cli::cli_h1("Machine setup")
 
-    cli_li("reduce the risk of accidentally committing sensitive files.")
-    cli_code("usethis::git_vaccinate()")
-    cli_text("\n")
+    cli::cli_li("reduce the risk of accidentally committing sensitive files.")
+    cli::cli_code("usethis::git_vaccinate()")
+    cli::cli_text("\n")
 
-    cli_li("use ssh instead of https when connecting to e.g. GitHub.")
-    cli_code("usethis::use_git_protocol('ssh')")
-    cli_text("\n")
+    cli::cli_li("use ssh instead of https when connecting to e.g. GitHub.")
+    cli::cli_code("usethis::use_git_protocol('ssh')")
+    cli::cli_text("\n")
 
-    cli_li("configures R to warn on partial matches.")
-    cli_code("usethis::use_partial_warnings()")
-    cli_text("\n")
+    cli::cli_li("configures R to warn on partial matches.")
+    cli::cli_code("usethis::use_partial_warnings()")
+    cli::cli_text("\n")
 
-    cli_li("configure Rstudio from the commandline.")
-    cli_code("install.packages('rstudio.prefs')")
-    cli_code("?rstudio.prefs::use_rstudio_prefs")
+    cli::cli_li("configure Rstudio from the commandline.")
+    cli::cli_code("install.packages('rstudio.prefs')")
+    cli::cli_code("?rstudio.prefs::use_rstudio_prefs")
   })
 }
 
@@ -43,59 +43,59 @@ setup_project <- function(path = NULL, repo = NULL) {
   path <- ifelse(is.null(path), "", paste0("'", path, "'"))
   repo <- ifelse(is.null(repo), "", paste0("'", repo, "'"))
 
-  cli({
+  cli::cli({
     ## machine setup
     setup_machine()
 
     ## new projects
-    cli_h1("New projects")
-    cli_li("create a new project.")
-    cli_code(paste0("usethis::create_project(", path, ")"))
-    cli_text("\n")
+    cli::cli_h1("New projects")
+    cli::cli_li("create a new project.")
+    cli::cli_code(paste0("usethis::create_project(", path, ")"))
+    cli::cli_text("\n")
 
     ### modules and packages
-    cli_h2("Adding modules")
-    cli_li("use a SpaDES module within the project.")
-    cli_code("use_module(path = 'modules', repo = '<user|org>/<repo>@<branch>')")
-    cli_text("\n")
+    cli::cli_h2("Adding modules")
+    cli::cli_li("use a SpaDES module within the project.")
+    cli::cli_code("use_module(path = 'modules', repo = '<user|org>/<repo>@<branch>')")
+    cli::cli_text("\n")
 
-    cli_li("Identify module package dependencies.")
-    cli_code(paste0("get_module_packages(path = 'modules')"))
-    cli_text("\n")
+    cli::cli_li("Identify module package dependencies.")
+    cli::cli_code(paste0("get_module_packages(path = 'modules')"))
+    cli::cli_text("\n")
 
     ### additional files
-    cli_h2("Additional files")
-    cli_li("use README with this project.")
-    cli_code("usethis::use_readme_md()")
-    cli_text("\n")
+    cli::cli_h2("Additional files")
+    cli::cli_li("use README with this project.")
+    cli::cli_code("usethis::use_readme_md()")
+    cli::cli_text("\n")
 
-    cli_li("use NEWS with this project.")
-    cli_code("usethis::use_news_md()")
-    cli_text("\n")
+    cli::cli_li("use NEWS with this project.")
+    cli::cli_code("usethis::use_news_md()")
+    cli::cli_text("\n")
 
-    cli_li("create DESCRIPTION file that captures package dependencies")
-    cli_code("description()")
-    cli_text("\n")
+    cli::cli_li("create DESCRIPTION file that captures package dependencies")
+    cli::cli_code("description()")
+    cli::cli_text("\n")
 
     ### git and github
-    cli_h2("Git and GitHub")
-    cli_li("use git with this project.")
-    cli_code("usethis::use_git()")
-    cli_text("\n")
+    cli::cli_h2("Git and GitHub")
+    cli::cli_li("use git with this project.")
+    cli::cli_code("usethis::use_git()")
+    cli::cli_text("\n")
 
-    cli_li("connect this project to a new GitHub repository.")
-    cli_code("usethis::use_github()")
+    cli::cli_li("connect this project to a new GitHub repository.")
+    cli::cli_code("usethis::use_github()")
 
     ### activate
-    cli_h2("Activate the project")
-    cli_li("activate the new project (will restart Rstudio).")
-    cli_code(paste0("usethis::proj_activate(", path, ")"))
-    cli_text("\n")
+    cli::cli_h2("Activate the project")
+    cli::cli_li("activate the new project (will restart Rstudio).")
+    cli::cli_code(paste0("usethis::proj_activate(", path, ")"))
+    cli::cli_text("\n")
 
     ## existing projects
-    cli_h1("Existing projects")
-    cli_code(paste0("usethis::create_from_github(", repo, ")"))
-    cli_li("")
-    cli_text("\n")
+    cli::cli_h1("Existing projects")
+    cli::cli_code(paste0("usethis::create_from_github(", repo, ")"))
+    cli::cli_li("")
+    cli::cli_text("\n")
   })
 }